Allow str() to be used on all types (#33)
* Move natives to their own files

* Change values to be returned as a malloc'd string rather than printed

* Malloc to calloc, and free output

* str() able to be used on all types. Tests for built in functions

* Fallback for unknown object type

* Remove test function
