+++ c/natives.h
@@ -1,0 +1,6 @@
+#ifndef dictu_natives_h
+#define dictu_natives_h
+
+void defineAllNatives();
+
+#endif //dictu_natives_h
