+++ c/object.h
@@ -180,7 +180,7 @@ ObjDict *initDict();
 
 ObjUpvalue *newUpvalue(Value *slot);
 
-void printObject(Value value);
+char *objectToString(Value value);
 
 static inline bool isObjType(Value value, ObjType type) {
     return IS_OBJ(value) && AS_OBJ(value)->type == type;
