+++ docs/docs/built-ins.md
@@ -11,12 +11,13 @@ nav_order: 9
 | print    | Prints arguments to the console                               | print(10, "test", nil, true);  |
 | input    | Returns user input as a string                                | input(); input("Input: ");     |
 | type     | Returns a string of the type of variable.                     | type(x);                       |
-| str      | Convert number to string value                                | str(10.2);                     |
+| str      | Convert a value to a string                                   | str(10.2); str([1, 2]); str({"hi": 10}) |
 | number   | Convert string to number value                                | number("10.2");                |
 | time     | Returns UNIX timestamp                                        | time();                        |
+| clock     | Returns number of clock ticks since the start of the program | clock();                        |
 | sleep    | Pauses execution for given time in seconds                    | sleep(10);                     |
 | assert   | Runtime error raised if condition doesn't evaluate to true    | assert(true);                  |
-| len      | Returns the length of a string/list                           | len("Hello"); len([1, 2]);     |
+| len      | Returns the length of a string or collection                  | len("Hello"); len([1, 2]); len({"test": 10});   |
 | min      | Returns the smallest number of the arguments                  | min(1, 2, 3); min([1, 2, 3]);  |
 | max      | Returns the largest number of the arguments                   | max(1, 2, 3); max([1, 2, 3]);  |
 | average  | Returns an average number of the arguments                    | average(1, 2, 3); average([1, 2, 3]); |
