+++ tests/builtins/len.du
@@ -1,0 +1,4 @@
+assert(len("123456789") == 9);
+assert(len([1, 2, 3, 4, 5]) == 5);
+assert(len([1, [2, 3], [4, 5, [6]]]) == 3);
+assert(len({"test": [1, 2], "hello": 10}) == 2);
