+++ tests/runTests.du
@@ -5,6 +5,7 @@ import "tests/dicts/import.du";
 import "tests/operators/import.du";
 import "tests/functions/import.du";
 import "tests/classes/import.du";
+import "tests/builtins/import.du";
 
 // If we got here no runtime errors were thrown, therefore all tests passed.
 print("All tests passed successfully!");
